Moved handler structures to a handler package
